backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
websockets==12.0
python-dateutil==2.8.2
pandas==2.1.4
numpy==1.26.2
apscheduler==3.10.4
pyyaml==6.0.1
requests==2.31.0
bcrypt==4.0.1